Added an `enableShallowLoading` option on Model.define(…) to allow loading documents without their associations with `[model].all({ shallow: true }, …)`
This is implemented with a `[model name]/_shallow` view that emits a copy of the document without the associations as a value. It is an error to call `[model].all({ shallow: true }, …)` on a model without `enableShallowLoading` since the required view would not be present.

FIXME: mark the objects as shallow to prevent saving them back to the db.